Desktop/llm-etl-cowboys/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pdfplumber==0.11.4
openai==1.51.0
pandas==2.2.2
python-dotenv==1.0.1
supabase==2.6.0
streamlit==1.37.1
plotly==5.24.0
modal==0.62.183
